ddns/app/requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.3
acme==0.1.1
argparse==1.2.1
cffi==1.4.2
cryptography==1.1.2
dnspython==1.12.0
enum34==1.1.2
funcsigs==0.4
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
jose==1.0.0
mock==1.3.0
ndg-httpsclient==0.4.0
pbr==1.8.1
pyOpenSSL==0.15.1
pyRFC3339==1.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pytz==2015.7
requests==2.9.1
rsa==3.2.3
six==1.10.0
wsgiref==0.1.2
